+++ app/src/main/java/com/xsolla/android/storesdkexample/StoreActivity.kt
@@ -49,6 +49,7 @@ import com.xsolla.android.storesdkexample.ui.vm.VmBalance
 import com.xsolla.android.storesdkexample.ui.vm.VmGooglePlay
 import com.xsolla.android.storesdkexample.ui.vm.VmProfile
 import com.xsolla.android.storesdkexample.ui.vm.base.ViewModelFactory
+import com.xsolla.android.appcore.utils.MiscUtils
 
 class StoreActivity : AppCompatActivity(R.layout.activity_store) {
 
@@ -100,6 +101,7 @@ class StoreActivity : AppCompatActivity(R.layout.activity_store) {
             val intent = XPayments.createIntentBuilder(this)
                 .accessToken(AccessToken(token))
                 .isSandbox(BuildConfig.IS_SANDBOX)
+                .setActivityType(MiscUtils.deduceXPaymentsActivityType(this))
                 .build()
             startActivityForResult(intent, RC_PAYSTATION)
         }