+++ xsolla-payments-sdk/src/main/java/com/xsolla/android/payments/ui/utils/CustomTabsHelper.kt
@@ -7,32 +7,41 @@ import androidx.browser.customtabs.CustomTabsClient
 import androidx.browser.customtabs.CustomTabsServiceConnection
 import androidx.browser.customtabs.CustomTabsSession
 
-class CustomTabsHelper(private val context: Context, private val payStation3WarmUpUrl: String, private val payStation4WarmUpUrl: String) {
-
+class CustomTabsHelper(
+    private val context: Context,
+    private val payStation3WarmUpUrl: String,
+    private val payStation4WarmUpUrl: String,
+    private val onCustomTabsSessionCreated: (customTabsSession: CustomTabsSession) -> Unit
+) {
+    companion object {
+        var IS_SUCCESSFULLY_INITIALIZED: Boolean = true
+    }
     private var customTabsSession: CustomTabsSession? = null
     private var mClient: CustomTabsClient? = null
 
     private var connection: CustomTabsServiceConnection? = object : CustomTabsServiceConnection() {
         override fun onCustomTabsServiceConnected(name: ComponentName, client: CustomTabsClient) {
-            if (client != null) {
-                mClient = client
+            mClient = client
+            if(client != null) {
                 client.warmup(0)
                 customTabsSession = client.newSession(null)
                 customTabsSession!!.mayLaunchUrl(Uri.parse(payStation3WarmUpUrl), null, null)
                 customTabsSession!!.mayLaunchUrl(Uri.parse(payStation4WarmUpUrl), null, null)
-
+                onCustomTabsSessionCreated(customTabsSession!!)
             }
-       }
+            IS_SUCCESSFULLY_INITIALIZED = mClient != null
+        }
 
         override fun onServiceDisconnected(name: ComponentName) {
             mClient = null
             customTabsSession = null
+            IS_SUCCESSFULLY_INITIALIZED = false
         }
     }
 
     fun bindCustomTabsService() {
-
         if (mClient != null) return
+
         val availableBrowsers = BrowserUtils.getAvailableCustomTabsBrowsers(context)
 
         if(availableBrowsers.isEmpty()) return